mysite/requirements.txt
djangorestframework
dj-rest-auth 
django-allauth 
firebase-admin 
django-cors-headers
django-rest-framework